requirements.txt
Flask-SQLAlchemy
numpy
pandas
psycopg2
python-dateutil
python-dotenv
requests
SQLAlchemy